refactor project folder structure and entry point
