--- index.js
@@ -1,3 +1,0 @@
-const functions = require("./console_app.js");
-
-functions.askForColor();
