+++ package.json
@@ -1,4 +1,5 @@
 {
+  "main": "src/index.js",
   "scripts": {
     "test": "jest"
   }
