+++ src/color-factories/color_factory.js
