+++ src/color-factories/hex_factory.js
