+++ src/color-factories/hsv_factory.js
