+++ src/color-factories/rgb1_factory.js
