+++ src/color-factories/rgb255_factory.js
