+++ src/index.js
@@ -1,0 +1,3 @@
+const { startConsoleApp } = require("./console-app/console_app");
+
+startConsoleApp();
