+++ src/tests/color_factory.test.js
@@ -1,6 +1,6 @@
 const { expect } = require("@jest/globals");
 
-const color_factory = require("./color_factory.js");
+const color_factory = require("../color-factories/color_factory");
 
 test('color factory has colorTypesEnum', () => {
   expect(color_factory).toHaveProperty('ColorTypes');
